Fix two big function bugs
The first is that we were double evaluating the arguments passed to
clojure functions (once in the evaluate function, and once in the `let`
form it invoked).

Second, is that we were using clojures let to create each function’s
local scope without compensating for a small semantic difference
between how clojure let bindings and function bindings work. See
comments inline.