Macros: add a very simple implementation
Clojure processes macros at compile time, of course, which we aren’t
doing. Clojure also identifies a function as a macro using metadata on
the var holding the function…but we don’t have vars or metadata.

So, we just store an `isMacro` property on function objects in place of
metadata, and evaluate functions with that property differently
(passing them their arguments unevaluated, and then evaluating the code
they return).

This doesn’t yet expose functionality for defining macros (and related
utilities like gensym).