Add defmacro macro
Note: in the code we generate, we include the setMacro function
directly, rather than a symbol that will resolve to it. I think this
would be bad practice in clojure (though I’m pretty sure it’s also
legal), but it makes sense here because our symbols aren’t namespaced
(yet), so returning a symbol would put us at too big a risk of our code
breaking accidentally if that symbol’s value was changed.