Refactor
The parse function doesn’t need to know the particulars of constructing
each data type from a series of items, esp. since this logic may need
to be reused elsewhere (e.g. in the `hash-map` function). So, instead,
move all that logic to “builder” objects that support a common
interface.